Moved some functions to the model
