+++ package.json
@@ -65,7 +65,7 @@
     "ts-jest": "^29.0.1",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.1",
-    "typescript": "^5.4.5",
+    "typescript": "^5.5.4",
     "webpack": "^5.91.0",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4",
