+++ src/models/VisualStyleModel/impl/cxVisualPropertyConverter.ts
@@ -6,8 +6,6 @@ import {
   NodeBorderLineType,
   NodeShapeType,
   VisualPropertyValueType,
-  HorizontalAlignType,
-  VerticalAlignType,
   VisibilityType,
   EdgeLineType,
   NodeLabelPositionType,
@@ -220,8 +218,8 @@ export const VPFontTypeConverter = (
 ): CXVisualPropertyConverter<FontType> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: CXFontFaceType): FontType =>
-      cxVPValue.FONT_FAMILY as FontType,
+    valueConverter: (cxVPValue: CXVisualPropertyValue): FontType =>
+      (cxVPValue as CXFontFaceType).FONT_FAMILY as FontType,
   }
 }
 
@@ -250,7 +248,8 @@ export const VPVisibilityTypeConverter = (
 ): CXVisualPropertyConverter<VisibilityType> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: VisibilityType): VisibilityType => cxVPValue,
+    valueConverter: (cxVPValue: CXVisualPropertyValue): VisibilityType =>
+      cxVPValue as VisibilityType,
   }
 }
 
@@ -259,7 +258,8 @@ export const VPEdgeLineTypeConverter = (
 ): CXVisualPropertyConverter<EdgeLineType> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: EdgeLineType): EdgeLineType => cxVPValue,
+    valueConverter: (cxVPValue: CXVisualPropertyValue): EdgeLineType =>
+      cxVPValue as EdgeLineType,
   }
 }
 
@@ -268,10 +268,10 @@ export const VPEdgeArrowShapeTypeConverter = (
 ): CXVisualPropertyConverter<EdgeArrowShapeType> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: EdgeArrowShapeType): EdgeArrowShapeType =>
+    valueConverter: (cxVPValue: CXVisualPropertyValue): EdgeArrowShapeType =>
       cxVPValue === EdgeArrowShapeType.Arrow
         ? EdgeArrowShapeType.Triangle
-        : cxVPValue,
+        : (cxVPValue as EdgeArrowShapeType),
   }
 }
 export const VPBooleanConverter = (
@@ -279,7 +279,8 @@ export const VPBooleanConverter = (
 ): CXVisualPropertyConverter<boolean> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: boolean): boolean => cxVPValue,
+    valueConverter: (cxVPValue: CXVisualPropertyValue): boolean =>
+      cxVPValue as boolean,
   }
 }
 
@@ -288,8 +289,8 @@ export const VPNodeLabelPositionConverter = (
 ): CXVisualPropertyConverter<NodeLabelPositionType> => {
   return {
     cxVPName,
-    valueConverter: (cxVPValue: NodeLabelPositionType): NodeLabelPositionType =>
-      cxVPValue,
+    valueConverter: (cxVPValue: CXVisualPropertyValue): NodeLabelPositionType =>
+      cxVPValue as NodeLabelPositionType,
   }
 }
