+++ src/models/tsconfig.json
@@ -1,10 +1,11 @@
 {
   "compilerOptions": {
-    "target": "es2020",
+    "target": "ES2020",
     "declaration": true,
     "emitDeclarationOnly": true,
     "skipLibCheck": true,
     "outDir": "./dist",
+    "rootDir": ".",
     "strict": true,
     "module": "NodeNext",
     "moduleResolution": "nodenext"
